hoàn thành các chức năng: Gửi phiếu thông báo kiểm tra y tế các nội dung kiểm tra cho phụ huynh xác nhận --> Chuẩn bị danh sách học sinh kiểm tra --> Thực hiện kiểm tra và ghi nhận kết quả

+++ src/main/java/com/medical/schoolMedical/repositories/ParentRepositoty.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ParentRepositoty extends JpaRepository<Parent,Integer> {
+public interface ParentRepositoty extends JpaRepository<Parent,Long> {
 }
