+++ src/main/java/com/medical/schoolMedical/repositories/UserRepository.java
@@ -5,11 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import org.springframework.stereotype.Repository;
 
-import java.util.BitSet;
-import java.util.Optional;
-
 @Repository
-public interface UserRepository extends JpaRepository<User,Integer> {
+public interface UserRepository extends JpaRepository<User,Long> {
     boolean existsByUsername(String username);
     User findByUsername(String username);
 }
